move rpmpython.h to more appropriate location
